pyproject.toml
[tool.poetry]
name = "gofannon"
version = "0.25.11"
description = "A collection of tools for LLMs"
authors = ["Trevor Grant <trevor.d.grant@gmail.com>"]
license = "ASFv2"
readme = "README.md"
packages = [{include = "gofannon"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.60.2"
requests = "^2.32.3"
GitPython = "^3.1.43"
python-dotenv = "^1.0.1"
jsonschema = "^4.23.0"
pygithub = "^2.6.1"

pytest = { version = "^8.3.4", optional = true }
langchain = { version = "^0.3.16", optional = true }
pydantic = { version = "^2.10.6", optional = true }
smolagents = { version = "^1.6.0", optional = true }
selenium = { version = "^4.10.0", optional = true }
google-api-python-client = { version = "^2.161.0", optional = true}
requests-mock = {version = "^1.12.1", optional = true}
boto3 = { version = "^1.34.97", optional = true }
jupyter = "^1.1.1"

[tool.poetry.extras]
testing = ["pytest","requests-mock"]
langchain = ["langchain", "pydantic"]
smolagents = ["smolagents"]
headless_browser = ["selenium"]
google = ["google-api-python-client"]
aws = ["boto3"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"